Moved Integrate SimpleInjector into Seaal.DependencyInjection.SimpleInjector
