+++ Seaal.DependencyInjection.SimpleInjector/Seaal.DependencyInjection.SimpleInjector.csproj
@@ -5,7 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.0.0" />
     <PackageReference Include="SimpleInjector" Version="4.0.11" />
+    <PackageReference Include="SimpleInjector.Integration.AspNetCore.Mvc" Version="4.0.11" />
   </ItemGroup>
 
   <ItemGroup>
